reorganizationand randomizing simple benching
