+++ code/Genomes/CMakeLists.txt
@@ -11,7 +11,8 @@ project( genomes )
 set(CMAKE_BUILD_TYPE Debug)
 
 # include source
-include_directories(${CMAKE_CURRENT_SOURCE_DIR}/../Tools)
+get_filename_component(PARENT_DIR ${CMAKE_CURRENT_SOURCE_DIR} DIRECTORY)
+include_directories(${PARENT_DIR}/Tools)
 
 # Optionally you can use file glob (uncomment the next line)
 file( GLOB_RECURSE GENOME_SOURCES *.cpp )
